+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Please add changes at the top. When releasing a version add a new header for tha
 
 ### master
 
+ - More than one type of aggregation can be selected for each attribute in one operation.
  - New _"Combine segmentations"_ operation makes it possible to create a segmentation
    by cities of residence _and_ age, for example.
  - New _"Segment by double attribute"_ and _"Segment by string attribute"_ operations